.dagger/pyproject.toml
[project]
name = "semantic-router-ci"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["dagger-io"]

[project.entry-points."dagger.mod"]
main_object = 'semantic_router_ci:SemanticRouter'

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }
